+++ src/program/ramsearch/RamWatchDetailed.h
@@ -33,25 +33,13 @@ public:
 
     T get_value()
     {
-        struct iovec local, remote;
+        update_addr();
+
+        if (!isValid)
+            return 0;
+
         T value = 0;
-        if (isPointer) {
-            address = base_address;
-            for (auto offset : pointer_offsets) {
-                uintptr_t next_address;
-                local.iov_base = static_cast<void*>(&next_address);
-                local.iov_len = sizeof(uintptr_t);
-                remote.iov_base = reinterpret_cast<void*>(address);
-                remote.iov_len = sizeof(uintptr_t);
-
-                isValid = (process_vm_readv(game_pid, &local, 1, &remote, 1, 0) == sizeof(uintptr_t));
-                if (!isValid)
-                    return value;
-
-                address = next_address + offset;
-            }
-        }
-
+        struct iovec local, remote;
         local.iov_base = static_cast<void*>(&value);
         local.iov_len = sizeof(T);
         remote.iov_base = reinterpret_cast<void*>(address);